requirements.txt
blinker==1.6.2
click==8.1.7
colorama==0.4.6
Flask==2.3.3
Flask-Cors==4.0.0
Flask-Email==1.4.4
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-WTF==1.1.1
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
psycopg2==2.9.7
python-decouple==3.8
python-dotenv==1.0.0
Werkzeug==2.3.7
WTForms==3.0.1
zipp==3.17.0
retrying==1.3.4